// src/app/core/api/generated/model/createPlanCommand.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanType } from './planType';


export interface CreatePlanCommand { 
    name?: string | null;
    description?: string | null;
    minDeductMinutes?: number;
    deductUnitMinutes?: number;
    planType?: PlanType;
    shiftIds?: Array<string> | null;
}