// src/app/core/api/generated/model/planDto.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanFactorProfileDto } from './planFactorProfileDto';


export interface PlanDto { 
    id?: string;
    name?: string | null;
    description?: string | null;
    isActive?: boolean;
    minDeductMinutes?: number;
    deductUnitMinutes?: number;
    factorProfiles?: Array<PlanFactorProfileDto> | null;
}